Survey usability lab upload, dropzone multiple upload support

+++ components/_shared/Dropzone/Dropzone.vue
@@ -12,11 +12,12 @@
     <span v-if="dz && dz.files.length" class="dropzone-delete" @click="clearItems">
       delete
     </span>
-    <div v-else class="dropzone-empty">
-      <IconFileImage />
-      <p class="dropzone-empty-text">
-        Upload a photo / logo
-      </p>
+    <div v-else-if="$slots.empty" class="dropzone-empty">
+      <slot name="empty" />
+    </div>
+    <slot name="custom-bg" />
+    <div class="dropzone-item-delete template">
+      <IconDelete />
     </div>
   </div>
 </template>
@@ -32,26 +33,61 @@ export default {
   props: {
     mutation: {
       type: String,
-      required: true
+      default: null
+    },
+    thumbnailWidth: {
+      type: Number,
+      default: null
+    },
+    thumbnailHeight: {
+      type: Number,
+      default: null
+    },
+    multiUpload: {
+      type: Boolean,
+      default: false
     }
   },
   data() {
     return {
       id: uuid.v4(),
-      dz: null
+      dz: null,
+      selected: []
     }
   },
   computed: {
     options() {
-      return {
+      const options = {
         url: `${process.env.baseUrl}/mediaFiles/input`,
         thumbnailMethod: 'contain',
         acceptedFiles: 'image/*',
-        thumbnailWidth: 400,
         maxFilesize: 5,
         dictDefaultMessage: '',
         dictResponseError: 'Request failed with status code {{statusCode}}',
         autoQueue: false
+      }
+      const { thumbnailWidth, thumbnailHeight } = this
+      if (thumbnailWidth) {
+        options.thumbnailWidth = thumbnailWidth
+      }
+      if (thumbnailHeight) {
+        options.thumbnailHeight = thumbnailHeight
+      }
+      return options
+    }
+  },
+  watch: {
+    selected(value) {
+      let result
+      if (this.multiUpload) {
+        result = value.map(item => item.id)
+      } else {
+        result = value.length ? value[0].id : null
+      }
+
+      this.$emit('change', result)
+      if (this.mutation) {
+        this.$store.commit(this.mutation, result)
       }
     }
   },
@@ -60,12 +96,19 @@ export default {
   },
   methods: {
     addedFile(newFile) {
-      if (this.dz.files.length > 1) {
-        const oldFiles = this.dz.files.filter(file => file.dataURL !== newFile.dataURL)
+      if (!this.multiUpload && this.dz.files.length > 1) {
+        const oldFiles = this.dz.files.filter(file => file.upload.uuid !== newFile.upload.uuid)
         oldFiles.forEach((oldFile) => {
           this.dz.removeFile(oldFile)
         })
       }
+      const deleteTemplate = document.querySelector('.dropzone-item-delete.template').cloneNode(true)
+      deleteTemplate.classList.remove('template')
+      deleteTemplate.addEventListener('click', () => {
+        this.dz.removeFile(newFile)
+        this.selected = this.selected.filter(item => item.dzId !== newFile.upload.uuid)
+      })
+      newFile.previewElement.appendChild(deleteTemplate)
       // do not ask why
       newFile.accepted = true
 
@@ -83,11 +126,14 @@ export default {
       xhr.setRequestHeader('Authorization', `Bearer ${this.$store.state.accessToken}`)
     },
     success(file, res) {
-      this.$store.commit(this.mutation, res.id)
+      this.selected.push({
+        dzId: file.upload.uuid,
+        id: res.id
+      })
     },
     clearItems() {
       this.dz.removeAllFiles(true)
-      this.$store.commit(this.mutation, null)
+      this.selected = []
     }
   }
 }
